MAINT: move _rng_html_rewrite to _util
